Clean up modin.experimental.cloud package API
