+++ modin/experimental/cloud/__init__.py
@@ -11,46 +11,46 @@
 # ANY KIND, either express or implied. See the License for the specific language
 # governing permissions and limitations under the License.
 
-from typing import NamedTuple
-
-class ClusterError(Exception):
-    """
-    Generic cluster operating exception
-    """
-
-    def __init__(self, *args, cause=None, **kw):
-        self.cause = cause
-        super().__init__(*args, **kw)
-
-
-class CannotSpawnCluster(ClusterError):
-    """
-    Raised when cluster cannot be spawned in the cloud
-    """
-
-
-class CannotDestroyCluster(ClusterError):
-    """
-    Raised when cluster cannot be destroyed in the cloud
-    """
-
-
-class ConnectionDetails(NamedTuple):
-    user_name: str = "modin"
-    key_file: str = None
-    address: str = None
-    port: int = 22
-
+from .base import ClusterError, CannotSpawnCluster, CannotDestroyCluster
 from .cluster import Provider, Cluster
-def cluster(provider: Provider,
-        project_name: str = None,
-        cluster_name: str = "modin-cluster",
-        worker_count: int = 4,
-        head_node_type: str = None,
-        worker_node_type: str = None,
-        spawner: str = 'rayscale') -> Cluster:
-    if spawner == 'rayscale':
+from .connection import Connection
+
+
+def cluster(
+    provider: Provider,
+    project_name: str = None,
+    cluster_name: str = "modin-cluster",
+    worker_count: int = 4,
+    head_node_type: str = None,
+    worker_node_type: str = None,
+    spawner: str = "rayscale",
+) -> Cluster:
+    if spawner == "rayscale":
         from .rayscale import RayCluster as Spawner
     else:
-        raise ValueError(f'Unknown spawner: {spawner}')
-    return Spawner(provider, project_name, cluster_name, worker_count, head_node_type, worker_node_type)
+        raise ValueError(f"Unknown spawner: {spawner}")
+    return Spawner(
+        provider,
+        project_name,
+        cluster_name,
+        worker_count,
+        head_node_type,
+        worker_node_type,
+    )
+
+
+def get_connection():
+    """
+    Returns an RPyC connection object to execute Python code remotely.
+    """
+    return Connection.get()
+
+
+__all__ = [
+    "ClusterError",
+    "CannotSpawnCluster",
+    "CannotDestroyCluster",
+    "Provider",
+    "Cluster",
+    "get_connection",
+]
